backend/requirements.txt
fastapi
uvicorn
chromadb
PyMuPDF
pydantic
reportlab
httpx
python-multipart
sqlalchemy
pyjwt
passlib[bcrypt]
bcrypt
